// src/api/generated/models/TenantResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eGovFrame 5.0 API
 * API Documentation for eGovFrame 5.0 Application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TenantResponse
 */
export interface TenantResponse {
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    tenantName?: string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    tenantDescription?: string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof TenantResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof TenantResponse
     */
    updatedDate?: Date;
}

/**
 * Check if a given object implements the TenantResponse interface.
 */
export function instanceOfTenantResponse(value: object): value is TenantResponse {
    return true;
}

export function TenantResponseFromJSON(json: any): TenantResponse {
    return TenantResponseFromJSONTyped(json, false);
}

export function TenantResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TenantResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'tenantId': json['tenantId'] == null ? undefined : json['tenantId'],
        'tenantName': json['tenantName'] == null ? undefined : json['tenantName'],
        'tenantDescription': json['tenantDescription'] == null ? undefined : json['tenantDescription'],
        'useAt': json['useAt'] == null ? undefined : json['useAt'],
        'createdDate': json['createdDate'] == null ? undefined : (new Date(json['createdDate'])),
        'updatedDate': json['updatedDate'] == null ? undefined : (new Date(json['updatedDate'])),
    };
}

export function TenantResponseToJSON(json: any): TenantResponse {
    return TenantResponseToJSONTyped(json, false);
}

export function TenantResponseToJSONTyped(value?: TenantResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'tenantId': value['tenantId'],
        'tenantName': value['tenantName'],
        'tenantDescription': value['tenantDescription'],
        'useAt': value['useAt'],
        'createdDate': value['createdDate'] == null ? undefined : ((value['createdDate']).toISOString()),
        'updatedDate': value['updatedDate'] == null ? undefined : ((value['updatedDate']).toISOString()),
    };
}